# apps/docs/static/openapi-corelink-v1.yaml
openapi: 3.1.0
info:
  title: CoreLink Customer & Privacy REST API
  version: 0.1.0
  summary: REST surface for CoreLink customer onboarding, privacy (DSR / DPA),
    billing webhooks, admin operations and health.
  description: |
    Hand-authored OpenAPI 3.1 contract for the **non-REAPI** surface of CoreLink.
    The high-volume CAS / AC / ByteStream / Capabilities surface is gRPC
    (Remote Execution API v2) and documented separately under
    `docs/reference/reapi/`.

    Endpoint canonical paths are sourced from the implementing crates' doc-comments:
    `corelink-signup`, `corelink-dpa-acceptance`, `corelink-dpa-versioning`,
    `corelink-tier-selection`, `corelink-pat`, `corelink-dsr`,
    `corelink-admin-api`,
    `corelink-enterprise-inquiry`, plus `apps/server::webhook` (Stripe).

    **Stability:** `v1` — public, semver-compatible. Breaking changes require a
    new major path prefix (`/v2/...`).
  termsOfService: https://corelink-docs.humangr.com/legal/terms
  contact:
    name: CoreLink Support
    email: support@humangr.com
    url: mailto:support@humangr.com
  license:
    name: Proprietary
    url: https://corelink-docs.humangr.com/legal/terms
servers:
  - url: https://corelink-api.humangr.com
    description: Production (Cloudflare Durable Object + container)
  - url: http://localhost:50051
    description: Local dev (axum container, PORT env default)
...
security:
  - BearerPAT: []
  - ClerkSessionCookie: []
paths:
...
  /bazel/v2/{instance}/findMissingBlobs:
    x-status: implemented
    x-implementing-handler: crates/corelink-container/src/routes/bazel_v2.rs (handle_find_missing)
    post:
      tags: [bazel-v2]
      summary: Find missing blobs (Bazel REAPI FindMissingBlobs POST).
      operationId: bazelFindMissingBlobs
      security:
        - BearerPAT: []
      parameters:
        - name: instance
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              description: REAPI FindMissingBlobsRequest (JSON-encoded proto).
      responses:
        "200":
          description: List of digests not present in the cache.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403":
          description: Cross-tenant attempt.
        "413":
          description: Batch too large.
components:
  securitySchemes:
    BearerPAT:
      type: http
      scheme: bearer
      bearerFormat: PAT
      description: |
        `Authorization: Bearer <PAT>` where PAT is a 32-byte UUIDv7-derived
        opaque token (BLAKE3-keyed hash in DB). Tier 1 of the canonical
        auth ladder per `corelink-pat`. Scopes encoded in PAT row.
    ClerkSessionCookie:
      type: apiKey
      in: cookie
      name: __session
      description: Clerk browser session cookie (alternative to PAT for the
        web onboarding flow).
    StripeSignature:
      type: apiKey
      in: header
      name: Stripe-Signature
      description: HMAC-SHA256 webhook signature; verified by
        `apps/server::webhook::stripe_webhook_handler` against the
        secret in `STRIPE_WEBHOOK_SECRET` env var with a 5-minute
        replay window.
...
  responses:
...
    Unauthorized:
      description: Missing or invalid credentials.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable machine-readable code from the canonical
                error taxonomy (`error_taxonomy.md`).
              examples: ["COR_ABUSE_DETECTED", "COR_DPA_REQUIRED",
                "COR_MFA_STEP_UP_REQUIRED", "COR_RATE_LIMITED"]
            message:
              type: string
              description: Human-readable explanation (locale-neutral; do
                NOT echo back PII).
            request_id:
              type: string
              format: uuid
              description: Server-assigned request id for support
                correlation.
            details:
              type: object
              additionalProperties: true
...